added CYGWIN (IA32 only) support added #ifdef HAVE_PRAGMA_WEAK logic around #pragma weak statements in mpi/c removed unused variable Host (run/TVSetUp.cc:70) fixed bug with wrong pointer to ATOMIC_UNLOCK() macros (mpi/c/mpi_errhandler_set.c line 52 and 55) cast unsigned to ssize_t for comparison (mem/FixedSharedMemPool.cc:71)
changes quickly reviewed with one harried Rob...Cheers, Mitch

+++ src/mem/FixedSharedMemPool.cc
@@ -68,7 +68,7 @@ void FixedSharedMemPool::init(ssize_t initialAllocation,
 
     // set minimum allocation size
     minAllocationSize_m = minAllocationSize;
-    if (minAllocationSize_m < getpagesize())
+    if (minAllocationSize_m < (ssize_t)getpagesize())
         minAllocationSize_m = getpagesize();
 
     // set nArrayElementsToAdd_m
